Replace auth callback methods
Removes authentication_reattempt_callback, which supports re-attempts at
authentication, with a single use basic_auth_callback. The latter will
only be invoked if either the username or password has not been set by
other means (configuration or in the connection URI).

This works around an bug introduced in Neo4j 3.1 (neo4j#8050), at the
expense of forcing re-attempts at authentication to occur in clients of
the driver.