Fix #8 - reset properly
This was a fairly deep logical bug issue the original code -- which
probably didn't matter before there were transactions.

The basic problem is that the reset method short circuits if the
connection is marked as failed -- but that is precisely when you want
to attempt a reset.

To be fair -- the ACK_FAILURE / RESET process of Bolt 1 might have
worked differently, so that the original logic made sense.