workaround unusual bug
Refering to messages.h/.c:
In libneo4j-client, message types (BEGIN, FAILURE, SUCCESS, etc.) are
represented by constant pointers to static structures of the form
{ .name, .struct_signature }
where the name is the Bolt all-caps name as a string constant, and
the signature is the 1-byte code that represents that msg in the
protocol.

Care is taken to ensure that these pointers do not move, so that a
type is tested as (e.g.)
  type == NEO4J_SUCCESS_MESSAGE
where the rhs is #define'd as the constant pointer value.

Strangely (to me), these pointers are not stable when using the
library in Perl XS functions (specifically, with the library in
Neo4j::Client and the XS in Neo4j::Bolt).

Nevertheless, the contents of the structures are ok, so that if you
check a type by looking at the name,
  strcmp(type->name, "SUCCESS")
you always get what you want (i.e., only the addresses seem to move
around, but the data is not stomped on).

To preserve the original code for standalone compilation, I've wrapped
each message type test in
doing the string check in each case if -DNEOCLIENT_BUILD.